backend/image-service/requirements.txt
fastapi
motor
python-multipart
pydantic
boto3              # For AWS S3
python-dotenv
Pillow             # For compression
aiofiles           # For file handling
pymongo
uvicorn[gunicorn]